requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.2
diff-match-patch==20230430
dj-database-url==2.2.0
Django==5.0.6
django-admin-inline-paginator==0.4.0
django-appconf==1.0.6
django-crispy-forms==2.2
django-crispy-formset-modal==0.12
django-debug-toolbar==4.4.2
django-extra-views==0.14.0
django-import-export==4.1.1
django-multiselectfield==0.1.13
django-select2==8.2.0
gunicorn==22.0.0
packaging==24.1
pillow==10.4.0
pip==24.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dotenv==1.0.1
setuptools==69.5.1
sqlparse==0.5.0
tablib==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
wheel==0.43.0
whitenoise==6.7.0